Extract multi-day scheduling helpers to shared module
The scheduling editor and the post-notification listing each
defined their own near-identical comma-list parsing/formatting
helpers. Moved parseSelectedValues, serializeSelectedValues, and
formatSelectedValues to assets/js/src/newsletters/scheduling/multi-day.ts
so both consumers share one implementation. formatSelectedValues
now reuses parseSelectedValues, which means listings show
selected days sorted numerically and de-duplicated -- matching the
storage path (which already sorts on write).

MAILPOET-8026
